Improve GUI, move GUI files into subpackage
Added a title to the Java GUI window, added a box around messages.
Tried to put profile pictures into circles but not very successful.
