+++ src/main/java/duke/Parser.java
@@ -9,6 +9,7 @@ import duke.tasks.Deadline;
 import duke.tasks.Event;
 import duke.tasks.Task;
 import duke.tasks.ToDo;
+import duke.ui.Ui;
 
 /**
  * Encapsulates the logic to parse and understand inputs by the user.
