+++ src/main/java/duke/ui/DialogBox.java
@@ -1,8 +1,9 @@
-package duke;
+package duke.ui;
 
 import java.io.IOException;
 import java.util.Collections;
 
+import duke.MainWindow;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
@@ -13,6 +14,9 @@ import javafx.scene.control.Label;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.HBox;
+import javafx.scene.paint.Color;
+import javafx.scene.shape.Circle;
+import javafx.scene.shape.StrokeType;
 
 
 public class DialogBox extends HBox {
@@ -20,6 +24,7 @@ public class DialogBox extends HBox {
     private Label dialog;
     @FXML
     private ImageView displayPicture;
+    private Circle clip;
 
     private DialogBox(String text, Image img) {
         try {
@@ -33,6 +38,11 @@ public class DialogBox extends HBox {
 
         dialog.setText(text);
         displayPicture.setImage(img);
+
+        clip = new Circle(displayPicture.getFitHeight()/2);
+        clip.setCenterX(displayPicture.getX() + displayPicture.getFitWidth()/2);
+        clip.setCenterY(displayPicture.getY() + displayPicture.getFitHeight()/2);
+        displayPicture.setClip(clip);
     }
 
     /**
