+++ src/main/resources/view/DialogBox.fxml
@@ -7,7 +7,8 @@
 
 <fx:root alignment="TOP_RIGHT" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefWidth="400.0" type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1">
   <children>
-    <Label fx:id="dialog" text="Label" wrapText="true" />
+    <Label fx:id="dialog" text="Label" wrapText="true" 
+      style="-fx-background-color: #72d4a9; -fx-background-radius: 10; -fx-padding: 10;" />
     <ImageView fx:id="displayPicture" fitHeight="99.0" fitWidth="99.0" pickOnBounds="true" preserveRatio="true" />
   </children>
   <padding>
